+++ data/src/config/context_menu.rs
@@ -1,9 +1,19 @@
 use serde::Deserialize;
 
-#[derive(Debug, Clone, Deserialize, Default)]
+#[derive(Debug, Clone, Deserialize)]
 #[serde(default)]
 pub struct ContextMenu {
     pub padding: Padding,
+    pub show_user_metadata: bool,
+}
+
+impl Default for ContextMenu {
+    fn default() -> Self {
+        Self {
+            padding: Padding::default(),
+            show_user_metadata: true,
+        }
+    }
 }
 
 #[derive(Debug, Copy, Clone, Deserialize)]